Agregar endpoints para Producto

+++ Models/Dtos/ProductoDto.cs
@@ -10,7 +10,7 @@ namespace Models.Dtos
         public string NombreProducto { get; set; }
         public string Categoria { get; set; }
         public string Marca { get; set; }
-        public double Precio { get; set; }
-        public double Costo { get; set; }
+        public decimal Precio { get; set; }
+        public decimal Costo { get; set; }
     }
 }
